add system that let users to earn points using the command plunder
the command plunder adds 2500 points to the user in the database.
user-checker.js inserts a new row into the db if user id does not exist
yet. this is ran for every command that reads/writes to the db.
points-updater.js is a generic helper function that adds points to the
specified user using user id.

cooldown on the command plunder is still needed to be implemented.